Move S2IAMValidatePrincipal gate to go/internal/gates.
Keeps the gate internal to the module and updates AWS, Azure, and GCP
verifier imports to use the shared internal package.
